Legg til placeholdere for oppgave 2 og 3 i workshop, og legg til ressurser som trengs
