+++ fasilitering/løsning/app/src/main/kotlin/no/jpro/kafkaworkshop/oppgave2/Producer.kt
@@ -1,16 +1,24 @@
 package no.jpro.kafkaworkshop.oppgave2
 
-import no.jpro.kafkaworkshop.Common.Message
-import no.jpro.kafkaworkshop.Common.objectMapper
-import no.jpro.kafkaworkshop.Common.topic
-import no.jpro.kafkaworkshop.Common.logger
+import com.fasterxml.jackson.module.kotlin.jacksonObjectMapper
+import no.jpro.kafkaworkshop.oppgave2.Common.logger
+import no.jpro.kafkaworkshop.oppgave2.Common.objectMapper
+import no.jpro.kafkaworkshop.oppgave2.Common.topic
 
 import org.apache.kafka.clients.CommonClientConfigs
 import org.apache.kafka.clients.producer.KafkaProducer
 import org.apache.kafka.clients.producer.ProducerConfig
 import org.apache.kafka.clients.producer.ProducerRecord
 import org.apache.kafka.common.serialization.StringSerializer
+import org.slf4j.LoggerFactory
 
+object Common {
+    val logger = LoggerFactory.getLogger("com.jpro.kafkaworkshop")
+    val objectMapper = jacksonObjectMapper()
+    const val topic = "kotlin_topic"
+
+    data class Message(val id: String, val value: String)
+}
 fun main() {
     val producerProperties = mapOf<String, Any>(
         ProducerConfig.BOOTSTRAP_SERVERS_CONFIG to "localhost:9092",
@@ -20,7 +28,7 @@ fun main() {
     )
 
     KafkaProducer<String, String>(producerProperties).use { producer ->
-        val message = Message(id = "1", value = "a value")
+        val message = Common.Message(id = "1", value = "a value")
         val jsonMessage: String = objectMapper.writeValueAsString(message)
         val record = ProducerRecord(topic, "key", jsonMessage)
