+++ workshop/app/build.gradle.kts
@@ -1,49 +1,28 @@
-/*
- * This file was generated by the Gradle 'init' task.
- *
- * This generated file contains a sample Kotlin application project to get you started.
- * For more details on building Java & JVM projects, please refer to https://docs.gradle.org/8.2.1/userguide/building_java_projects.html in the Gradle documentation.
- */
-
 plugins {
-    // Apply the org.jetbrains.kotlin.jvm Plugin to add support for Kotlin.
-    id("org.jetbrains.kotlin.jvm") version "1.8.20"
-
-    // Apply the application plugin to add support for building a CLI application in Java.
-    application
+    java
+    kotlin("jvm") version "1.8.20"
 }
 
+group = "no.jpro.kafkaworkshop"
+version = "1.0-SNAPSHOT"
+
 repositories {
-    // Use Maven Central for resolving dependencies.
     mavenCentral()
 }
 
 dependencies {
-    // Use the Kotlin JUnit 5 integration.
-    testImplementation("org.jetbrains.kotlin:kotlin-test-junit5")
+    implementation("org.apache.kafka:kafka-streams:3.4.0")
+    implementation("org.apache.kafka:kafka-clients:3.4.0")
+    implementation("ch.qos.logback:logback-classic:1.2.3")
+    implementation("ch.qos.logback:logback-core:1.2.3")
+    implementation ("com.fasterxml.jackson.core:jackson-databind:2.15.2")
+    implementation("com.fasterxml.jackson.module:jackson-module-kotlin:2.15.2")
+    implementation(kotlin("stdlib"))
 
-    // Use the JUnit 5 integration.
-    testImplementation("org.junit.jupiter:junit-jupiter-engine:5.9.2")
-
-    testRuntimeOnly("org.junit.platform:junit-platform-launcher")
-
-    // This dependency is used by the application.
-    implementation("com.google.guava:guava:31.1-jre")
 }
 
-// Apply a specific Java toolchain to ease working on different environments.
-java {
-    toolchain {
-        languageVersion.set(JavaLanguageVersion.of(17))
+tasks.withType<org.jetbrains.kotlin.gradle.tasks.KotlinCompile>().configureEach {
+    kotlinOptions {
+        jvmTarget = "18"
     }
-}
-
-application {
-    // Define the main class for the application.
-    mainClass.set("no.jpro.kafkaworkshop.AppKt")
-}
-
-tasks.named<Test>("test") {
-    // Use JUnit Platform for unit tests.
-    useJUnitPlatform()
 }
