+++ workshop/app/src/main/kotlin/kafkaworkshop/oppgave2/Producer.kt
@@ -1,0 +1,5 @@
+package no.jpro.kafkaworkshop.oppgave2
+
+fun main() {
+    // TODO: Oppgave 2
+}
