+++ workshop/app/src/main/kotlin/kafkaworkshop/oppgave3/Consumer.kt
@@ -1,0 +1,5 @@
+package no.jpro.kafkaworkshop.oppgave3
+
+fun main() {
+    // TODO: Oppgave 3
+}
