--- workshop/app/src/main/kotlin/no/jpro/kafkaworkshop/App.kt
@@ -1,15 +1,0 @@
-/*
- * This Kotlin source file was generated by the Gradle 'init' task.
- */
-package no.jpro.kafkaworkshop
-
-class App {
-    val greeting: String
-        get() {
-            return "Hello World!"
-        }
-}
-
-fun main() {
-    println(App().greeting)
-}
