--- workshop/app/src/test/kotlin/no/jpro/kafkaworkshop/AppTest.kt
@@ -1,14 +1,0 @@
-/*
- * This Kotlin source file was generated by the Gradle 'init' task.
- */
-package no.jpro.kafkaworkshop
-
-import kotlin.test.Test
-import kotlin.test.assertNotNull
-
-class AppTest {
-    @Test fun appHasAGreeting() {
-        val classUnderTest = App()
-        assertNotNull(classUnderTest.greeting, "app should have a greeting")
-    }
-}
